Only compress when there is compressed file

+++ m4_ext/compress.bash
@@ -11,14 +11,18 @@ DIR=$(dirname "${FILE}")
 if [[ "$EXT" = "jpg" ]] || [[ "$EXT" = "jpeg" ]]; then
 	new_name="$BASE"_comp.jpg
 
-	cp $FILE $new_name
-	jpegoptim $new_name > /dev/null
+	if ! [[ -f "$new_name" ]]; then
+		cp $FILE $new_name
+		jpegoptim $new_name > /dev/null
+	fi
 	printf "$new_name"
 elif [[ "$EXT" = "png" ]]; then
 	new_name="$BASE"_comp.png
 
-	cp $FILE $new_name
-	optipng $new_name > /dev/null
+	if ! [[ -f "$new_name" ]]; then
+		cp $FILE $new_name
+		optipng $new_name > /dev/null
+	fi
 	printf "$new_name"
 else # Unsupported extesnio file is not processed
 	echo $FILE
